# src/main/resources/customerChat-openapi_1.yaml
openapi: "3.1.0"
info:
  title: "customerChat API"
  description: "customerChat API"
  version: "1.0.0"
servers:
  - url: "https://customerChat"
paths:
  /api/chat/{id}:
    post:
      summary: "POST api/chat/{id}"
      operationId: "appendMessage"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "userId"
          in: "header"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              type: "string"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      summary: "DELETE api/chat/{id}"
      operationId: "deleteChatSession"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
  /api/chat:
    post:
      summary: "POST api/chat"
      operationId: "createChatSession"
      parameters:
        - name: "customerId"
          in: "header"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatSession"
  /api/chat/{id}/messages:
    get:
      summary: "GET api/chat/{id}/messages"
      operationId: "getMessages"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Message"
components:
  schemas:
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        username:
          type: "string"
      required:
        - "username"
    ChatSession:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        customer:
          $ref: "#/components/schemas/User"
        agent:
          $ref: "#/components/schemas/User"
        messages:
          type: "array"
          items:
            $ref: "#/components/schemas/Message"
      required:
        - "customer"
        - "agent"
        - "messages"
    Message:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        sender:
          $ref: "#/components/schemas/User"
        content:
          type: "string"
        timestamp:
          type: "string"
          format: "date-time"
        session:
          $ref: "#/components/schemas/ChatSession"
      required:
        - "sender"
        - "content"
        - "timestamp"
        - "session"